+++ mpu9250/__init__.py
@@ -1,7 +1,8 @@
 """Driver for the InvenSense MPU-9250 9-axis IMU (accel + gyro + AK8963 magnetometer) over I2C."""
 
-from .data import MPUCalData, MPUData
-from .demo import DemoMPU
+from .bmp280 import BMP280
+from .data import BaroData, MPUCalData, MPUData
+from .demo import DemoBaro, DemoMPU
 from .driver import MPU9250, HardwareMismatchError
 from .fusion import MahonyAHRS
 from .ranges import AccelRange, GyroRange, LPF
@@ -10,8 +11,10 @@ from .ticker import TickerThread
 
 __all__ = [
     'MPU9250',
+    'BMP280',
     'MPUData',
     'MPUCalData',
+    'BaroData',
     'AccelRange',
     'GyroRange',
     'LPF',
@@ -20,4 +23,5 @@ __all__ = [
     'MahonyAHRS',
     'Recorder',
     'DemoMPU',
+    'DemoBaro',
 ]